[mdf][all] Документы под целочисленный мир
Документы приведены в соответствие с двумя предыдущими изменениями:

docs/01_Cosmology.md — мягкость почвы при эрозии теперь `1 - каменистость`
(утрамбованности в мире нет), и добавлено следствие, которого раньше не
было сказано: на горах мало что растёт не по отдельному правилу, а по той
же сухости камня.

docs/08_Plants.md — таблица черт короче на шесть строк; пригодность клетки
описана одной влажностью; ниша осталась одна, и это записано в
ограничения модели честно, как потеря.

docs/09_Animals.md — таблица черт без пищеварения и с целыми диапазонами;
здоровье одно на все беды; раздел о вытаптывании переписан в раздел о том,
почему его больше нет.

README — новый раздел "Целые числа" со шкалой и тремя приёмами вместо
дробей; все примеры конфигурации, протокола и файла мира пересчитаны;
слои и клавиши перенумерованы.

Версии подняты: протокол 18 -> 19, формат мира 1 -> 2, приложение
0.23.0 -> 0.24.0. Формат мира — не из-за нового поля, а потому что прежние
числа сменили и смысл, и шкалу: "growth": 0.6 версии 1 и "growth": 600
версии 2 — не одно и то же, и молча прочитать первое как второе значило бы
открыть мир, которого не было. Файлы версии 1 отвергаются с ошибкой.

Заодно исправлено найденное проверкой: невыбранный запас шага рос без
конца у того, кто быстрее клетки за тик. Шагов от этого больше не
становилось, но число в теле уезжало в бесконечность — дробным оно уезжало
туда же, просто этого не было видно.

+++ core/src/generation/GrassSeeding.cpp
@@ -110,7 +110,7 @@ void seedGrass(World& world, const PlantParams& params, unsigned seed) {
         // у PlantSystem: тонет не тот, под кем есть вода, а тот, под кем
         // её больше, чем он переносит.
         const auto* water = world.registry().try_get<const WaterComponent>(terrain);
-        const float waterDepth = water != nullptr ? water->depth : 0.0f;
+        const int waterDepth = water != nullptr ? water->depth : 0;
 
         const std::size_t speciesIndex =
             static_cast<std::size_t>(randomUnit(state) * static_cast<float>(species.size())) % species.size();
